+++ contracts/LiquidityManager.sol
@@ -55,7 +55,7 @@ abstract contract LiquidityManager is IERC721Receiver, Ownable {
         return uint160(sqrtRatio * (2 ** 48)); // Scale the result by 2^48
     }
 
-    function _createLiquilityPool(address token0, address token1, uint24 fee) internal returns (address pool) {
+    function _createLiquidityPool(address token0, address token1, uint24 fee) internal returns (address pool) {
         require(token0 < token1, "token0 > token1");
 
         pool = uniswapV3Factory.getPool(token0, token1, fee);
@@ -66,14 +66,28 @@ abstract contract LiquidityManager is IERC721Receiver, Ownable {
         IUniswapV3Pool(pool).increaseObservationCardinalityNext(MIN_OBSERVATION_CARDINALITY);
     }
 
-    function _addLiquidity(
-        address token0,
-        uint256 amount0Desired,
-        address token1,
-        uint256 amount1Desired,
+    function _mintLiquidity(
+        address tokenAddress,
+        uint256 tokenAmount,
+        uint256 assetAmount,
         address recipient
-    ) internal returns (address pool, uint256 tokenId, uint128 liquidity, uint256 amount0, uint256 amount1) {
-        pool = _createLiquilityPool(token0, token1, UNISWAP_FEE);
+    ) internal returns (address pool, uint256 tokenId, uint128 liquidity, uint256 actualTokenAmount, uint256 actualAssetAmount) {
+        address token0;
+        address token1;
+        uint256 amount0Desired;
+        uint256 amount1Desired;
+        if (tokenAddress < address(WETH)) {
+            token0 = tokenAddress;
+            token1 = address(WETH);
+            amount0Desired = tokenAmount;
+            amount1Desired = assetAmount;
+        } else {
+            token1 = tokenAddress;
+            token0 = address(WETH);
+            amount1Desired = tokenAmount;
+            amount0Desired = assetAmount;
+        }
+        pool = _createLiquidityPool(token0, token1, UNISWAP_FEE);
 
         uint160 sqrtPriceX96 = getSqrtPriceX96(amount1Desired, amount0Desired);
 
@@ -96,7 +110,26 @@ abstract contract LiquidityManager is IERC721Receiver, Ownable {
             deadline: block.timestamp
         });
 
-        (tokenId, liquidity, amount0, amount1) = nonfungiblePositionManager.mint(params);
+        (tokenId, liquidity, actualTokenAmount, actualAssetAmount) = nonfungiblePositionManager.mint(params);
+        if (tokenAddress >= address(WETH)) {
+            (actualTokenAmount, actualAssetAmount) = (actualAssetAmount, actualTokenAmount);
+        }
+    }
+
+    function _increaseLiquidity(
+        uint256 tokenId,
+        uint256 amount0Desired,
+        uint256 amount1Desired
+    ) internal returns (uint128 liquidity, uint256 amount0, uint256 amount1) {
+        INonfungiblePositionManager.IncreaseLiquidityParams memory params = INonfungiblePositionManager.IncreaseLiquidityParams({
+            tokenId: tokenId,
+            amount0Desired: amount0Desired,
+            amount1Desired: amount1Desired,
+            amount0Min: 0,
+            amount1Min: 0,
+            deadline: block.timestamp
+        });
+        (liquidity, amount0, amount1) = nonfungiblePositionManager.increaseLiquidity(params);
     }
 
     function getSpotSqrtPriceX96(address tokenAddress) public view returns (uint160) {