+++ contracts/TokenFactory.sol
@@ -29,6 +29,7 @@ contract TokenFactory is ReentrancyGuard, LiquidityManager {
 
     mapping(address => address) public tokensCreators;
     mapping(address => address) public tokensPools;
+    mapping(address => uint256) public liquidityPositionTokenIds;
 
     // Events
     event TokenCreated(address indexed token, string name, string symbol, string uri, address creator, uint256 competitionId, uint256 timestamp);
@@ -36,6 +37,8 @@ contract TokenFactory is ReentrancyGuard, LiquidityManager {
     event NewCompetitionStarted(uint256 competitionId, uint256 timestamp);
 
     event TokenBuy(address indexed token, uint256 amount0In, uint256 amount0Out, uint256 fee, uint256 timestamp);
+
+    event TokenMinted(address indexed token, uint256 assetAmount, uint256 tokenAmount, uint256 timestamp);
 
     event TokenSell(address indexed token, uint256 amount0In, uint256 amount0Out, uint256 fee, uint256 timestamp);
 
@@ -58,8 +61,8 @@ contract TokenFactory is ReentrancyGuard, LiquidityManager {
         address sender,
         uint256 tokenId,
         uint128 liquidity,
-        uint256 amount0,
-        uint256 amount1,
+        uint256 actualTokenAmount,
+        uint256 actualAssetAmount,
         uint256 timestamp
     );
 
@@ -145,7 +148,7 @@ contract TokenFactory is ReentrancyGuard, LiquidityManager {
     }
 
     function _buyReceivedAmount(address tokenAddress, uint256 paymentAmount) public view returns (uint256 tokenAmount) {
-        (uint256 paymentWithoutFee, ) = _getCollateralAmountAndFee(paymentAmount);
+        (uint256 paymentWithoutFee,) = _getCollateralAmountAndFee(paymentAmount);
         return _getBuyTokenAmount(tokenAddress, paymentWithoutFee);
     }
 
@@ -177,7 +180,7 @@ contract TokenFactory is ReentrancyGuard, LiquidityManager {
         token.burn(from, tokenAmount);
         if (to != address(this)) {
             //slither-disable-next-line arbitrary-send-eth
-            (bool success, ) = to.call{value: paymentAmountWithoutFee}(new bytes(0));
+            (bool success,) = to.call{value: paymentAmountWithoutFee}(new bytes(0));
             require(success, "ETH send failed");
         }
         emit TokenSell(tokenAddress, tokenAmount, paymentAmountWithoutFee, fee, block.timestamp);
@@ -212,7 +215,7 @@ contract TokenFactory is ReentrancyGuard, LiquidityManager {
             if (winnerTokenAddress == tokenAddress) {
                 collateralWithoutFee += collateralById[competitionId][tokenAddress];
             } else {
-                (uint256 paymentWithoutFee, ) = _getCollateralAmountAndFee(collateralById[competitionId][tokenAddress]);
+                (uint256 paymentWithoutFee,) = _getCollateralAmountAndFee(collateralById[competitionId][tokenAddress]);
                 collateralWithoutFee += paymentWithoutFee;
             }
         }
@@ -246,12 +249,11 @@ contract TokenFactory is ReentrancyGuard, LiquidityManager {
             mintAmount = (currentCollateral * numTokensPerEther) / 1e18;
             Token(tokenAddress).mint(address(this), mintAmount);
         }
-        (address pool, uint256 tokenId, uint128 liquidity, uint256 amount0, uint256 amount1) = tokenAddress < address(WETH)
-            ? _addLiquidity(tokenAddress, mintAmount, address(WETH), currentCollateral, address(this))
-            : _addLiquidity(address(WETH), currentCollateral, tokenAddress, mintAmount, address(this));
+        (address pool, uint256 tokenId, uint128 liquidity, uint256 actualTokenAmount, uint256 actualAssetAmount) = _mintLiquidity(tokenAddress, mintAmount, currentCollateral, address(this));
 
         tokensPools[tokenAddress] = pool;
-        emit WinnerLiquidityAdded(tokenAddress, tokensCreators[tokenAddress], pool, msg.sender, tokenId, liquidity, amount0, amount1, block.timestamp);
+        liquidityPositionTokenIds[tokenAddress] = tokenId;
+        emit WinnerLiquidityAdded(tokenAddress, tokensCreators[tokenAddress], pool, msg.sender, tokenId, liquidity, actualTokenAmount, actualAssetAmount, block.timestamp);
     }
 
     function burnTokenAndMintWinner(address tokenAddress) external nonReentrant {
@@ -267,13 +269,19 @@ contract TokenFactory is ReentrancyGuard, LiquidityManager {
         uint256 netUserCollateral = userCollateral - fee;
         uint256 mintAmount = getMintAmountPostPublish(netUserCollateral, tokenAddress);
         Token(winnerToken).mint(msg.sender, mintAmount);
+        Token(winnerToken).mint(address(this), mintAmount);
+        _increaseLiquidity(liquidityPositionTokenIds[tokenAddress],
+            tokenAddress < address(WETH) ? mintAmount : netUserCollateral,
+            tokenAddress < address(WETH) ? netUserCollateral : mintAmount
+        );
+        emit TokenMinted(tokenAddress, netUserCollateral, mintAmount, block.timestamp);
         feeAccumulated += fee;
         emit BurnTokenAndMintWinner(msg.sender, tokenAddress, winnerToken, burnedAmount, mintAmount, fee, block.timestamp);
     }
 
     function withdrawFee() external onlyOwner {
         uint256 feeWithdrawable = feeAccumulated - feeWithdrawn > address(this).balance ? address(this).balance : feeAccumulated - feeWithdrawn;
-        (bool success, ) = owner().call{value: feeWithdrawable}("");
+        (bool success,) = owner().call{value: feeWithdrawable}("");
         require(success, "transfer failed");
         feeWithdrawn += feeWithdrawable;
     }